requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-functions==1.13.3
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
langchain==0.0.123
loguru==0.6.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pinecone-client==2.2.1
pydantic==1.10.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
yarl==1.8.2
